[Refactoring] decompose the implementations of different metrics into several files (#1161)
* refactor metrics
* add UT for refactored metrics
